requirements.txt
python-telegram-bot==22.3
python-dotenv==1.0.0
aiofiles==24.1.0
aiohttp==3.9.1
schedule==1.2.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3